Moved the dynamicactioninfostore to services
