+++ StreamDeckPlugin/Events/DynamicActionInfoChanged.cs
@@ -1,5 +1,4 @@
 ﻿using StreamDeckPlugin.Services;
-using StreamDeckPlugin.Utils;
 using System;
 
 namespace StreamDeckPlugin.Events {
